feat(reader): Delete Wallpaper in the pop-up
The in-book menu could already delete the wallpaper the lock screen last
showed, behind its own confirmation. It is now an action like any other, so it
can be ticked into the reader pop-up or bound to a button.

popupItems widened from 16 bits to 32. Delete Wallpaper is the first pop-up
member past bit 15, which was the whole ceiling before. No migration: the mask
persists as a plain number and every existing bit keeps its position.

Availability is the same test Hold Wallpaper uses, and the delete itself is the
menu's existing path, confirmation and card removal included.

+++ src/CrossPointSettings.cpp
@@ -330,16 +330,19 @@ bool CrossPointSettings::fromJson(JsonVariantConst doc) {
   // trimmed to POPUP_ITEM_MAX so a file claiming twenty rows cannot build a pop-up
   // taller than the panel.
   {
-    uint16_t storedPopupItems = doc["popupItems"] | (uint16_t)0;
-    uint16_t validMask = 0;
-    for (const uint8_t fn : POPUP_ITEM_FUNCTIONS) validMask |= static_cast<uint16_t>(1u << fn);
+    // Widened to 32 bits when Delete Wallpaper became the first pop-up row past bit 15.
+    // No migration: the mask persists as a plain number and every existing bit keeps its
+    // position, so a file written by an older build reads back unchanged.
+    uint32_t storedPopupItems = doc["popupItems"] | (uint32_t)0;
+    uint32_t validMask = 0;
+    for (const uint8_t fn : POPUP_ITEM_FUNCTIONS) validMask |= static_cast<uint32_t>(1u << fn);
     storedPopupItems &= validMask;
     popupItems = 0;
     uint8_t kept = 0;
     for (const uint8_t fn : POPUP_ITEM_FUNCTIONS) {
       if (!((storedPopupItems >> fn) & 1u)) continue;
       if (kept >= POPUP_ITEM_MAX) break;
-      popupItems |= static_cast<uint16_t>(1u << fn);
+      popupItems |= static_cast<uint32_t>(1u << fn);
       kept++;
     }
     if (popupItems != storedPopupItems) needsResave = true;
